Add hdf writer scan test into test_hdf_writer
